Random Generation Update
Random maze generation is now implemented! It has the functionality of supporting any n x m grid, but I haven't added a way to specify the parameters in the game yet. Will do that next update. Also working on a "labyrinth" mode that currently does not work if you try to play it.